Error function
The optional feature in which DADA calculates error rates from the
matrix of substitutions it infers now runs on a passed in function
parameter, err_function. err_function takes dada$trans_out, and returns
an error rate matrix. This is then used as the $err for the next loop
when self_consist = TRUE, or just reported as $err_out if not
self_consist looping.

Additionaly, documentation and naming consistency was improved.

isBimera is now provided as part of the package.

All dependence on seqinr was removed.

Time to merge back into main branch.

+++ src/cluster.cpp
@@ -65,7 +65,7 @@ Raw *raw_qual_new(char *seq, double *qual, int reads) {
   strcpy(raw->seq, seq);
   raw->qual = NULL;
   if(qual) {
-    raw->qual = (double *) malloc(seqlen * sizeof(double)); //E
+    raw->qual = (float *) malloc(seqlen * sizeof(float)); //E
     if (raw->qual == NULL)  Rcpp::stop("Memory allocation failed!\n");
     for(int i=0;i<seqlen;i++) { raw->qual[i] = qual[i]; }
   } else {
